copy button, label to managerViewController #51
버튼, 라벨을 관리자모드 viewController에 복사하고 연결시킴
